// src/vehicles/enums/document.enum.ts
export enum Role {
  ADMIN = "ADMIN",
  CONTROLLER = "CONTROLLER",
  OPERATOR = "OPERATOR",
  FINANCE = "FINANCE",
  CUSTOMER = "CUSTOMER",
  DRIVER = "DRIVER",
}

export enum STATUS {
  ACTIVE = "ACTIVE",
}

export enum VehicleStatus {
  ACTIVE = "ACTIVE",
  DISABLED = "DISABLED",
  SUSPENDED = "SUSPENDED",
}